reafactor(run): refactor creation of command runE to unify execution of all types of commands
